refactor(sdk-core): refactor common methods in TSS coins
Ticket: BG-58214
